native_posix: Refactor cmd line options
For each driver/model with command line options.
Move its code for the board file into the driver/model file.
